requirements.txt
streamlit
python-dotenv
pytesseract
pdf2image
pymupdf
groq
